Move std::ops to separate file
