+++ hoomd-linear-algebra/src/matrix.rs
@@ -1,12 +1,12 @@
 // Copyright (c) 2024-2025 The Regents of the University of Michigan.
 // Part of hoomd-rs, released under the BSD 3-Clause License.
 
+/// ``std::ops`` implementations for [`Matrix`]
+mod ops;
+
 pub use crate::diagonal::DiagonalMatrix;
 
-use std::{
-    fmt,
-    ops::{Add, AddAssign, Index, IndexMut, Mul, MulAssign, Neg, Sub, SubAssign},
-};
+use std::fmt;
 
 /// A lightweight representation of a diagonal matrix.
 use crate::{Diagonal, GeneralMatrix, Invertible, MatMul, QuadraticForm, SquareMatrix};
@@ -23,38 +23,6 @@ pub type Matrix22 = Matrix<2, 2>;
 pub type Matrix33 = Matrix<3, 3>;
 /// A 4x4 matrix, allocated on the stack.
 pub type Matrix44 = Matrix<4, 4>;
-/// Index the rows and columns of a [`Matrix`]
-///
-/// Indices for [`Matrix`] types are zero-indexed and reflect the indexing pattern of
-/// the underlying data. This results in the pattern `(row, column)`, which mirrors the
-/// behavior of Numpy and similar array languages.
-///
-/// # Examples
-/// ```
-/// use hoomd_linear_algebra::matrix::Matrix;
-/// let rows = [[1.0, 2.0], [3.0, 4.0], [5.0, 6.0]];
-/// let mat = Matrix { rows };
-/// assert_eq!(mat[(0, 1)], rows[0][1]);
-/// assert_eq!(mat[(2, 1)], 6.0);
-/// assert_eq!(mat[(1, 1)], 4.0);
-/// // Out-of-bounds: would panic!
-/// // mat[(3, 0)];
-/// ```
-impl<const N: usize, const M: usize> Index<(usize, usize)> for Matrix<N, M> {
-    type Output = f64;
-    #[inline]
-    fn index(&self, index: (usize, usize)) -> &f64 {
-        let (i, j) = index;
-        &self.rows[i][j]
-    }
-}
-impl<const N: usize, const M: usize> IndexMut<(usize, usize)> for Matrix<N, M> {
-    #[inline]
-    fn index_mut(&mut self, index: (usize, usize)) -> &mut f64 {
-        let (i, j) = index;
-        &mut self.rows[i][j]
-    }
-}
 
 impl<const N: usize, const M: usize> GeneralMatrix for Matrix<N, M> {
     #[inline]
@@ -525,116 +493,6 @@ impl<const N: usize> QuadraticForm for Matrix<N, N> {
     }
 }
 
-/// Compute the elementwise scalar multiplication of a [`Matrix`]
-///
-/// # Examples
-/// ```
-/// use hoomd_linear_algebra::{GeneralMatrix, matrix::Matrix22};
-/// let matrix = Matrix22::full(2.0);
-/// let scalar = 2.0;
-/// assert_eq!(matrix * scalar, matrix + matrix);
-/// ```
-impl<const N: usize, const M: usize> Mul<f64> for Matrix<N, M> {
-    type Output = Self;
-
-    #[inline]
-    fn mul(self, rhs: f64) -> Self {
-        self.map_elementwise(|x| x * rhs)
-    }
-}
-/// Multiply an `f64` scalar by a [`Matrix`] rhs.
-///
-/// # Examples
-/// ```
-/// use hoomd_linear_algebra::{GeneralMatrix, matrix::Matrix22};
-/// let matrix = Matrix22::full(2.0);
-/// let scalar = 3.0;
-/// assert_eq!(scalar * matrix, matrix * scalar);
-/// ```
-impl<const N: usize, const M: usize> Mul<Matrix<N, M>> for f64 {
-    type Output = Matrix<N, M>;
-
-    #[inline]
-    fn mul(self, rhs: Self::Output) -> Self::Output {
-        rhs.map_elementwise(|x| x * self)
-    }
-}
-
-/// Compute the elementwise, in-place scalar multiplication of a [`Matrix`]
-///
-/// # Examples
-/// ```
-/// use hoomd_linear_algebra::{GeneralMatrix, matrix::Matrix22};
-/// let mut matrix = Matrix22::full(2.0);
-/// let matrix_copy = matrix.clone();
-/// matrix *= 3.0;
-/// assert_eq!(matrix, matrix_copy * 3.0);
-/// ```
-impl<const N: usize, const M: usize> MulAssign<f64> for Matrix<N, M> {
-    #[inline]
-    fn mul_assign(&mut self, rhs: f64) {
-        self.iter_flat_mut().for_each(|x| *x *= rhs);
-    }
-}
-
-/// Compute the elementwise negation of a [`Matrix`].
-///
-/// # Examples
-/// ```
-/// use hoomd_linear_algebra::{GeneralMatrix, matrix::Matrix22};
-/// let matrix = Matrix22::full(5.0);
-/// assert_eq!(-matrix, Matrix22::zeros() - matrix);
-/// ```
-impl<const N: usize, const M: usize> Neg for Matrix<N, M> {
-    type Output = Self;
-
-    #[inline]
-    fn neg(self) -> Self {
-        self.map_elementwise(f64::neg)
-    }
-}
-
-impl<const N: usize, const M: usize> Add<Self> for Matrix<N, M> {
-    type Output = Self;
-
-    #[inline]
-    fn add(self, rhs: Self) -> Self {
-        Self {
-            rows: std::array::from_fn(|i| {
-                std::array::from_fn(|j| self.rows[i][j] + rhs.rows[i][j])
-            }),
-        }
-    }
-}
-impl<const N: usize, const M: usize> AddAssign for Matrix<N, M> {
-    #[inline]
-    fn add_assign(&mut self, rhs: Self) {
-        self.iter_flat_mut()
-            .zip(rhs.iter_flat())
-            .for_each(|(x, r)| *x += r);
-    }
-}
-impl<const N: usize, const M: usize> Sub<Self> for Matrix<N, M> {
-    type Output = Self;
-
-    #[inline]
-    fn sub(self, rhs: Self) -> Self {
-        Self {
-            rows: std::array::from_fn(|i| {
-                std::array::from_fn(|j| self.rows[i][j] - rhs.rows[i][j])
-            }),
-        }
-    }
-}
-impl<const N: usize, const M: usize> SubAssign for Matrix<N, M> {
-    #[inline]
-    fn sub_assign(&mut self, rhs: Self) {
-        self.iter_flat_mut()
-            .zip(rhs.iter_flat())
-            .for_each(|(x, r)| *x -= r);
-    }
-}
-
 impl Invertible for Matrix<2, 2> {
     /// Compute the inverse of a matrix. Will be `None` if the matrix is not invertible.
     ///
@@ -780,9 +638,7 @@ impl Matrix<2, 2> {
 
 /// Macro to generate impls for a given row size `N` and multiple column sizes `M`.
 macro_rules! impl_copy_for_m {
-    ($N:literal, $($M:literal),+) => {
-        $(#[doc(hidden)]impl Copy for Matrix<$N, $M> {})+
-    };
+    ($N:literal, $($M:literal),+) => { $(#[doc(hidden)]impl Copy for Matrix<$N, $M> {})+ };
 }
 /// Implement Copy for matrices of an input size `N`, `M`
 macro_rules! impl_copy_for_n_m {
@@ -795,10 +651,13 @@ impl<const N: usize> Diagonal for [f64; N] {}
 
 #[cfg(test)]
 mod tests {
-    use std::fmt::Debug;
+    use std::{
+        fmt::Debug,
+        ops::{Index, Mul},
+    };
 
     use super::*;
-    use crate::matrix::{Matrix, Matrix22};
+    use crate::matrix::{Matrix, Matrix22, Matrix33, Matrix44};
     use approx::{assert_relative_eq, assert_ulps_eq, ulps_eq};
     use faer::Mat;
     use rstest::rstest;
@@ -840,7 +699,7 @@ mod tests {
     }
     fn assert_diags_ulps_eq<const N: usize, T: Diagonal>(
         m0: &T,
-        m1: &impl Index<usize, Output = f64>,
+        m1: &impl std::ops::Index<usize, Output = f64>,
     ) {
         for i in 0..N {
             assert_ulps_eq!(m0[i], m1[i], epsilon = EPS);
@@ -1051,117 +910,6 @@ mod tests {
     }
 
     #[test]
-    fn test_matrix_add_2x2() {
-        let a_rows = [[1.0, 2.0], [3.0, 4.0]];
-        let b_rows = [[5.0, 6.0], [7.0, 8.0]];
-
-        let a = Matrix::<2, 2> { rows: a_rows };
-        let b = Matrix::<2, 2> { rows: b_rows };
-        let faer_a = fill_faer(a_rows);
-        let faer_b = fill_faer(b_rows);
-
-        let custom_sum = a + b;
-        let faer_sum = faer_a + faer_b;
-
-        assert_matrixes_ulps_eq::<2, 2, _, _>(&custom_sum, &faer_sum);
-        assert_matrixes_ulps_eq::<2, 2, _, _>(&custom_sum, &faer_sum);
-    }
-
-    #[test]
-    fn test_matrix_add_2x3() {
-        let a_rows = [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]];
-        let b_rows = [[7.0, 8.0, 9.0], [10.0, 11.0, 12.0]];
-        let a = Matrix::<2, 3> { rows: a_rows };
-        let b = Matrix::<2, 3> { rows: b_rows };
-        let faer_a = fill_faer(a_rows);
-        let faer_b = fill_faer(b_rows);
-
-        let custom_sum = a + b;
-        let faer_sum = faer_a + faer_b;
-
-        assert_matrixes_ulps_eq::<2, 3, _, _>(&custom_sum, &faer_sum);
-    }
-
-    #[test]
-    fn test_matrix_sub_2x2() {
-        let a_rows = [[1.0, 2.0], [3.0, 4.0]];
-        let b_rows = [[5.0, 6.0], [7.0, 8.0]];
-        let a = Matrix::<2, 2> { rows: a_rows };
-        let b = Matrix::<2, 2> { rows: b_rows };
-        let faer_a = fill_faer(a_rows);
-        let faer_b = fill_faer(b_rows);
-
-        let custom_diff = a - b;
-        let faer_diff = faer_a - faer_b;
-
-        assert_matrixes_ulps_eq::<2, 2, _, _>(&custom_diff, &faer_diff);
-    }
-
-    #[test]
-    fn test_matrix_sub_2x3() {
-        let a_rows = [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]];
-        let b_rows = [[7.0, 8.0, 9.0], [10.0, 11.0, 12.0]];
-        let a = Matrix::<2, 3> { rows: a_rows };
-        let b = Matrix::<2, 3> { rows: b_rows };
-        let faer_a = fill_faer(a_rows);
-        let faer_b = fill_faer(b_rows);
-
-        let custom_diff = a - b;
-        let faer_diff = faer_a - faer_b;
-
-        assert_matrixes_ulps_eq::<2, 3, _, _>(&custom_diff, &faer_diff);
-    }
-
-    #[rstest(
-        rows,
-        case([[1.0, -2.0], [3.0, 4.0]]),
-        case([[0.0, 0.0], [0.0, 0.0]])
-    )]
-    fn test_matrix_neg_2x2(rows: [[f64; 2]; 2]) {
-        let matrix = Matrix::<2, 2> { rows };
-        let faer_matrix = fill_faer(rows);
-
-        let custom_neg = -matrix;
-        let faer_neg = -faer_matrix;
-
-        assert_matrixes_ulps_eq::<2, 2, _, _>(&custom_neg, &faer_neg);
-    }
-
-    #[rstest]
-    #[case([[1.0, 2.0], [3.0, 4.0]], 5.0)]
-    #[case([[1.0, 2.0], [3.0, 4.0]], -1.0)]
-    #[case([[1.0, 2.0], [3.0, 4.0]], 0.0)]
-    fn test_matrix_scalar_mul_2x2(#[case] rows: [[f64; 2]; 2], #[case] scalar: f64) {
-        let matrix = Matrix::<2, 2> { rows };
-        let faer_matrix = fill_faer(rows);
-
-        let custom_prod = matrix * scalar;
-        let faer_prod = faer_matrix * scalar;
-
-        assert_matrixes_ulps_eq::<2, 2, _, _>(&custom_prod, &faer_prod);
-    }
-
-    #[test]
-    fn test_indexing() {
-        // Matrix
-        let mat = Matrix::<2, 3> {
-            rows: [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]],
-        };
-        assert_eq!(mat[(0, 2)], 3.0);
-        assert_eq!(mat[(1, 1)], 5.0);
-    }
-
-    #[test]
-    fn test_mut_indexing() {
-        let mut mat = Matrix::<2, 2>::zeros();
-        mat[(0, 1)] = 99.0;
-        mat[(1, 0)] = -5.5;
-        assert_eq!(mat[(0, 1)], 99.0);
-        assert_eq!(mat[(1, 0)], -5.5);
-        assert_eq!(mat[(1, 1)], 0.0);
-    }
-
-    #[test]
     fn test_general_matrix_methods() {
         // Matrix
         let zeros = Matrix::<2, 3>::zeros();
@@ -1267,77 +1015,5 @@ mod tests {
         let identity = Matrix44::identity();
 
         assert_matrixes_ulps_eq::<4, 4, _, _>(&product, &identity);
-    }
-    #[rstest]
-    #[case(
-        [[1.0, 2.0], [ 3.0, 4.0], [5.0, 6.0]],
-        [[7.0, 8.0], [9.0, 10.0], [11.0, 12.0]],
-    )]
-    #[case(
-        [[1.0, 0.0], [0.0, 1.0], [1.0, 1.0]],
-        [[2.0, 3.0], [ 4.0, 5.0], [6.0, 7.0]],
-    )]
-    #[case(
-        [[1.0],[ 2.0]],
-        [[3.0], [4.0]],
-    )]
-    #[case(
-        [[1.0, 2.0], [3.0, 4.0], [1.0, 1.0]],
-        [[1.0, 0.0], [0.0, 1.0], [1.0, 1.0]],
-    )]
-    fn test_add_assign<const M: usize, const N: usize>(
-        #[case] a_rows: [[f64; M]; N],
-        #[case] b_rows: [[f64; M]; N],
-    ) {
-        let mut a = Matrix { rows: a_rows };
-        let b = Matrix { rows: b_rows };
-        let c = a.clone() + b.clone();
-
-        a += b;
-        assert_eq!(a, c);
-    }
-    #[rstest]
-    #[case(
-        [[1.0, 2.0], [ 3.0, 4.0], [5.0, 6.0]], 0.0
-    )]
-    #[case(
-        [[1.0, 0.0], [0.0, 1.0], [1.0, 1.0]], -91.0
-    )]
-    #[case(
-        [[1.0],[ 2.0]], 33.3
-    )]
-    #[case(
-        [[1.0, 2.0], [3.0, 4.0], [1.0, 1.0]], 84.0
-    )]
-    fn test_mul_assign<const M: usize, const N: usize>(
-        #[case] a_rows: [[f64; M]; N],
-        #[case] x: f64,
-    ) {
-        let mut a = Matrix { rows: a_rows };
-        let c = a.clone() * x;
-
-        a *= x;
-        assert_eq!(a, c);
-    }
-
-    #[rstest]
-    #[case(
-        [[1.0, 2.0], [ 3.0, 4.0], [5.0, 6.0]], 0.0
-    )]
-    #[case(
-        [[1.0, 0.0], [0.0, 1.0], [1.0, 1.0]], -91.0
-    )]
-    #[case(
-        [[1.0],[ 2.0]], 33.3
-    )]
-    #[case(
-        [[1.0, 2.0], [3.0, 4.0], [1.0, 1.0]], 84.0
-    )]
-    fn test_mul_left<const M: usize, const N: usize>(
-        #[case] a_rows: [[f64; M]; N],
-        #[case] x: f64,
-    ) {
-        let a = Matrix { rows: a_rows };
-        assert_eq!(a.clone() * x, x * a);
     }
 }
